fix: extract urlBase64ToUint8Array to separate file to fix recurring CI failure
Bun's Linux transpiler intermittently fails to parse push.ts (which uses
browser-only PushManager/serviceWorker APIs), causing all exports to
become unavailable. Moving the pure utility to base64.ts means tests
never need to import from push.ts for this function.
