+++ frontend/src/lib/push.ts
@@ -6,17 +6,6 @@ export function isPushSupported(): boolean {
   );
 }
 
-export function urlBase64ToUint8Array(base64String: string): Uint8Array {
-  const padding = "=".repeat((4 - (base64String.length % 4)) % 4);
-  const base64 = (base64String + padding).replace(/-/g, "+").replace(/_/g, "/");
-  const rawData = atob(base64);
-  const outputArray = new Uint8Array(rawData.length);
-  for (let i = 0; i < rawData.length; i++) {
-    outputArray[i] = rawData.charCodeAt(i);
-  }
-  return outputArray;
-}
-
 export async function subscribeToPush(
   vapidPublicKey: string
 ): Promise<{ endpoint: string; p256dh: string; auth: string }> {
